Add a simple alternative "client" in pure-python
It allows other components that depend on swh-graph to run their tests
without depending on the WebGraph process itself.

+++ swh/graph/tests/conftest.py
@@ -1,3 +1,9 @@
+# Copyright (C) 2019-2021  The Software Heritage developers
+# See the AUTHORS file at the top-level directory of this distribution
+# License: GNU General Public License version 3, or any later version
+# See top-level LICENSE file for more information
+
+import csv
 import multiprocessing
 from pathlib import Path
 
@@ -7,6 +13,7 @@ import pytest
 from swh.graph.backend import Backend
 from swh.graph.client import RemoteGraphClient
 from swh.graph.graph import load as graph_load
+from swh.graph.naive_client import NaiveClient
 from swh.graph.server.app import make_app
 
 SWH_GRAPH_TESTS_ROOT = Path(__file__).parents[0]
@@ -33,16 +40,23 @@ class GraphServerProcess(multiprocessing.Process):
             self.q.put(e)
 
 
-@pytest.fixture(scope="module")
-def graph_client():
-    queue = multiprocessing.Queue()
-    server = GraphServerProcess(queue)
-    server.start()
-    res = queue.get()
-    if isinstance(res, Exception):
-        raise res
-    yield RemoteGraphClient(str(res))
-    server.terminate()
+@pytest.fixture(scope="module", params=["remote", "naive"])
+def graph_client(request):
+    if request.param == "remote":
+        queue = multiprocessing.Queue()
+        server = GraphServerProcess(queue)
+        server.start()
+        res = queue.get()
+        if isinstance(res, Exception):
+            raise res
+        yield RemoteGraphClient(str(res))
+        server.terminate()
+    else:
+        with open(SWH_GRAPH_TESTS_ROOT / "dataset/example.nodes.csv") as fd:
+            nodes = [node for (node,) in csv.reader(fd, delimiter=" ")]
+        with open(SWH_GRAPH_TESTS_ROOT / "dataset/example.edges.csv") as fd:
+            edges = list(csv.reader(fd, delimiter=" "))
+        yield NaiveClient(nodes=nodes, edges=edges)
 
 
 @pytest.fixture(scope="module")
